requirements.txt
airtable-python-wrapper==0.15.1
boto3==1.16.59
botocore==1.19.59
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
shortuuid==1.0.1
six==1.15.0
urllib3==1.26.2
